requirements.txt
altair==4.2.2
astropy==5.2.2
astroquery==0.4.6
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cmake==3.26.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
decorator==5.1.1
einops==0.6.1
entrypoints==0.4
filelock==3.9.0
fonttools==4.39.3
gitdb==4.0.10
GitPython==3.1.31
html5lib==1.1
idna==3.4
importlib-metadata==6.6.0
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keyring==23.13.1
kiwisolver==1.4.4
lit==16.0.2
llvmlite==0.39.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.2.1
networkx==3.0
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pandas==1.5.3
Pillow==9.5.0
protobuf==3.20.3
pyarrow==11.0.0
pycparser==2.21
pydeck==0.8.1b0
pyerfa==2.0.0.3
Pygments==2.15.1
Pympler==1.0.1
pynndescent==0.5.10
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyvo==1.4.1
PyYAML==6.0
requests==2.28.1
rich==13.3.4
scikit-learn==1.2.2
scipy==1.10.1
SecretStorage==3.3.3
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
streamlit==1.21.0
streamlit-image-coordinates==0.1.3
streamlit-space==0.1.5
streamlit-toggle==0.1.3
sympy==1.11.1
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tornado==6.3.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3
umap-learn==0.5.3
urllib3==1.26.13
validators==0.20.0
watchdog==3.0.0
webencodings==0.5.1
zipp==3.15.0
